+++ src/cpp/xpansion_interfaces/include/antares-xpansion/xpansion_interfaces/Clock.h
@@ -9,7 +9,10 @@ class Clock {
   Clock() = default;
   virtual ~Clock() = default;
 
-  virtual std::time_t getTime();
+  virtual std::time_t getTime() {
+    std::time_t beginTime = std::time(nullptr);
+    return beginTime;
+  }
 };
 
 inline void localtime_platform(const std::time_t &time_p,
@@ -21,5 +24,15 @@ inline void localtime_platform(const std::time_t &time_p,
 #endif
 }
 namespace clock_utils {
-std::string timeToStr(const std::time_t &time_p);
+inline std::string timeToStr(const std::time_t &time_p) {
+  struct tm local_time;
+  localtime_platform(time_p, local_time);
+  // localtime_r(&time_p, &local_time); // Compliant
+  const char *FORMAT = "%d-%m-%Y %H:%M:%S";
+  char buffer_l[100];
+  strftime(buffer_l, sizeof(buffer_l), FORMAT, &local_time);
+  std::string strTime_l(buffer_l);
+
+  return strTime_l;
+}
 }
